refactor: Using junban as workflow management for active learning tasks
