+++ biocentral_server/active_learning/pipelines/__init__.py
@@ -1,0 +1,3 @@
+from .protein_screening import al_screening_pipeline
+
+__all__ = ["al_screening_pipeline"]
