+++ biocentral_server/active_learning/pipelines/protein_screening/__init__.py
@@ -1,0 +1,3 @@
+from .screening_pipeline import al_screening_pipeline
+
+__all__ = ["al_screening_pipeline"]
