+++ pyproject.toml
@@ -32,6 +32,7 @@ dependencies = [
     "vespag @ git+https://github.com/biocentral/VespaG.git@onnx-deployment",
     "tmbed @ git+https://github.com/biocentral/TMbed.git@onnx-deployment",
     "apscheduler>=3.11.2",
+    "junban>=1.0.2",
 ]
 
 [dependency-groups]
